+++ src/myNotesProvider.ts
@@ -1,5 +1,7 @@
-import { Note } from '@hackmd/api/dist/type';
 import * as vscode from 'vscode';
+
+import { Note } from '@hackmd/api/dist/type';
+
 import { API } from './api';
 import { recordUsage } from './store';